Fix for PR 1367

+++ include/etl/type_lookup.h
@@ -73,7 +73,7 @@ namespace etl
 
 #if ETL_USING_CPP11 && !defined(ETL_TYPE_SELECT_FORCE_CPP03_IMPLEMENTATION)
   //***************************************************************************
-  // type_id_lookup 
+  // type_id_lookup
   //***************************************************************************
   template <typename... TTypes>
   struct type_id_lookup
@@ -84,21 +84,21 @@ namespace etl
     struct nulltype {};
 
     // For N type pairs.
-    template <int Id, typename T1, typename... TRest>
+    template <size_t Id, typename T1, typename... TRest>
     struct type_from_id_helper
     {
       using type = typename etl::conditional<Id == T1::Id,
-                                             typename T1::type,
-                                             typename type_from_id_helper<Id, TRest...>::type>::type;
+        typename T1::type,
+        typename type_from_id_helper<Id, TRest...>::type>::type;
     };
 
     // Specialisation for 1 type pair.
-    template <int Id, typename T1>
+    template <size_t Id, typename T1>
     struct type_from_id_helper<Id, T1>
     {
       using type = typename etl::conditional<Id == T1::Id,
-                                             typename T1::type,
-                                             nulltype>::type;
+        typename T1::type,
+        nulltype>::type;
     };
 
   public:
@@ -159,7 +159,7 @@ namespace etl
     {
       return get_id_from_type<T>();
     }
-     
+
     //************************************
     template <typename T>
     static unsigned int get_id_from_type()
@@ -183,16 +183,16 @@ namespace etl
     struct type_from_type_helper
     {
       using type = typename etl::conditional<etl::is_same<T, typename T1::type1>::value,
-                                             typename T1::type2,
-                                             typename type_from_type_helper<T, TRest...>::type>::type;
+        typename T1::type2,
+        typename type_from_type_helper<T, TRest...>::type>::type;
     };
 
     template <typename T, typename T1>
     struct type_from_type_helper<T, T1>
     {
       using type = typename etl::conditional<etl::is_same<T, typename T1::type1>::value,
-                                             typename T1::type2,
-                                             nulltype>::type;
+        typename T1::type2,
+        nulltype>::type;
     };
 
   public:
